rename package `web` to `controller`
